Rename Interaction to MessageRespondingInteraction
