+++ core/src/main/kotlin/behavior/interaction/EphemeralInteractionResponseBehavior.kt
@@ -1,11 +1,10 @@
 package dev.kord.core.behavior.interaction
 
-import dev.kord.common.annotation.KordUnsafe
 import dev.kord.common.entity.Snowflake
 import dev.kord.core.Kord
 
 /**
- * The behavior of a ephemeral [Discord Interaction Response](https://discord.com/developers/docs/interactions/slash-commands#interaction-response)
+ * The behavior of a ephemeral [Discord MessageRespondingInteraction Response](https://discord.com/developers/docs/interactions/slash-commands#interaction-response)
  * This response is visible to *only* to the user who made the interaction.
  */
