+++ core/src/main/kotlin/behavior/interaction/PublicInteractionResponseBehavior.kt
@@ -6,7 +6,7 @@ import dev.kord.rest.request.RestRequestException
 
 
 /**
- * The behavior of a public [Discord Interaction Response](https://discord.com/developers/docs/interactions/slash-commands#interaction-response)
+ * The behavior of a public [Discord MessageRespondingInteraction Response](https://discord.com/developers/docs/interactions/slash-commands#interaction-response)
  * This response is visible to all users in the channel.
  */
