+++ core/src/main/kotlin/event/interaction/InteractionCreate.kt
@@ -1,8 +1,8 @@
 package dev.kord.core.event.interaction
 
-import dev.kord.core.entity.interaction.Interaction
+import dev.kord.core.entity.interaction.MessageRespondingInteraction
 import dev.kord.core.event.Event
 
 public sealed interface InteractionCreateEvent : Event {
-    public val interaction: Interaction
+    public val interaction: MessageRespondingInteraction
 }
